Create GenreConfiguration
