+++ src/Domain/Entities/Genre.cs
@@ -6,7 +6,6 @@ namespace Domain.Entities
     {
         public int Id { get; set; }
 
-        [Required]
         public string Name { get; set; }
 
         public string? Description { get; set; }
