+++ src/Infrastructure/Data/ApplicationDbContext.cs
@@ -15,57 +15,13 @@ namespace Infrastructure.Data
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
-            modelBuilder.Entity<Genre>()
-                .Property(g => g.Id)
-                .ValueGeneratedOnAdd();
-
             modelBuilder.Entity<Platform>()
                 .Property(p => p.Id)
                 .ValueGeneratedOnAdd();
 
-            modelBuilder.Entity<Genre>().HasData(CreateGenreDataSeed());
             modelBuilder.Entity<Platform>().HasData(CreatePlatformDataSeed());
 
             base.OnModelCreating(modelBuilder);
-        }
-
-        private Genre[] CreateGenreDataSeed()
-        {
-            Genre[] newGenres = [
-
-                new Genre()
-                {
-                    Id = 1,
-                    Name = "None",
-                    Description = "Sin género"
-                },
-                new Genre()
-                {
-                    Id = 2,
-                    Name = "RTS",
-                    Description = "Real Time Strategy"
-                },
-                new Genre
-                {
-                    Id = 3,
-                    Name = "Sports",
-                    Description = null
-                },
-                new Genre
-                {
-                    Id = 4,
-                    Name = "Fighting",
-                    Description = null
-                },
-                new Genre
-                {
-                    Id = 5,
-                    Name = "Racing",
-                    Description = null
-                },
-            ];
-
-            return newGenres;
         }
 
         private Platform[] CreatePlatformDataSeed()
